requirements.txt
nginx
postgres
nodejs
